Move legacy actions closer to stores
